fix(controller): move to server/controller
